refactor(@deepnote/cli): import integration loading from @deepnote/database-integrations directly
Drop the backward-compat re-export shims left behind by the integration
loading extraction and have every CLI consumer import from
@deepnote/database-integrations (browser-safe root) or its /node subpath
directly.

- Delete the 9 re-export shims (constants, utils/{api,dotenv,env-var-refs},
  integrations/{fetch-integrations,inject-integration-env-vars,
  integrations-file-schemas,merge-integrations,parse-integrations}) and the
  integrations barrel
- Remove the re-export blocks from commands/integrations.ts and validate.ts
- Repoint all consumers, dynamic import() calls, and the generate-json-schemas
  script
- Retarget vitest mocks to the package root and /node specifiers
