+++ packages/cli/src/cli.ts
@@ -1,3 +1,9 @@
+import {
+  DEEPNOTE_TOKEN_ENV,
+  DEFAULT_API_URL,
+  DEFAULT_ENV_FILE,
+  DEFAULT_INTEGRATIONS_FILE,
+} from '@deepnote/database-integrations'
 import chalk from 'chalk'
 import { Command } from 'commander'
 // Note: We keep 'chalk' import for:
@@ -10,7 +16,7 @@ import { createDagDownstreamAction, createDagShowAction, createDagVarsAction } f
 import { createDiffAction } from './commands/diff'
 import { createInspectAction } from './commands/inspect'
 import { createInstallSkillsAction } from './commands/install-skills'
-import { createIntegrationsPullAction, DEFAULT_API_URL } from './commands/integrations'
+import { createIntegrationsPullAction } from './commands/integrations'
 import { createIntegrationsAddAction } from './commands/integrations/add-integration'
 import { createIntegrationsEditAction } from './commands/integrations/edit-integration'
 import { createLintAction } from './commands/lint'
@@ -19,7 +25,6 @@ import { createRunAction } from './commands/run'
 import { createStatsAction } from './commands/stats'
 import { createValidateAction } from './commands/validate'
 import { generateCompletionScript } from './completions'
-import { DEEPNOTE_TOKEN_ENV, DEFAULT_ENV_FILE, DEFAULT_INTEGRATIONS_FILE } from './constants'
 import { ExitCode } from './exit-codes'
 import { getChalk, getOutputConfig, OUTPUT_FORMATS, output, setOutputConfig, shouldDisableColor } from './output'
 import { createFormatValidator, JSON_LLM_RESOLUTION, TOON_LLM_RESOLUTION } from './utils/format-validator'
