+++ packages/cli/src/commands/integrations/tests/add-integration.spanner.test.ts
@@ -115,7 +115,7 @@ describe('add-integration spanner', () => {
     // Verify multiline values survive the updateDotEnv → readDotEnv roundtrip.
     // screen.type() cannot inject literal newlines (they act as Enter/submit),
     // so we write a multiline service account value directly and verify it parses back.
-    const { updateDotEnv, readDotEnv } = await import('../../../utils/dotenv')
+    const { updateDotEnv, readDotEnv } = await import('@deepnote/database-integrations/node')
     const multilineServiceAccount = 'line-1-of-service-account\nline-2-of-service-account\nline-3-of-service-account'
     await updateDotEnv(envFilePath, {
       AAAAAAAA_BBBB_CCCC_DDDD_EEEEEEEEEEEE__SERVICE_ACCOUNT: multilineServiceAccount,
