+++ packages/cli/src/commands/run.test.ts
@@ -3,13 +3,16 @@ import os from 'node:os'
 import { join } from 'node:path'
 import { deserializeDeepnoteFile } from '@deepnote/blocks'
 import type { DatabaseIntegrationConfig } from '@deepnote/database-integrations'
+import {
+  ApiError,
+  type ApiIntegration,
+  DEEPNOTE_TOKEN_ENV,
+  DEFAULT_API_URL,
+  DEFAULT_INTEGRATIONS_FILE,
+} from '@deepnote/database-integrations'
 import { Command } from 'commander'
 import { afterEach, beforeEach, describe, expect, it, type Mock, type MockedFunction, vi } from 'vitest'
-import { DEEPNOTE_TOKEN_ENV, DEFAULT_INTEGRATIONS_FILE } from '../constants'
-import type { ApiIntegration } from '../integrations/fetch-integrations'
-import { ApiError } from '../utils/api'
 import type { saveExecutionSnapshot } from '../utils/output-persistence'
-import { DEFAULT_API_URL } from './integrations'
 
 // Create mock engine functions
 const mockStart = vi.fn()
@@ -53,19 +56,13 @@ vi.mock('@deepnote/reactivity', () => {
   }
 })
 
-// Mock integrations module for testing integration validation
+// Mock integration parsing + env-var injection (now in @deepnote/database-integrations/node)
 const mockParseIntegrationsFile = vi.fn()
-vi.mock('../integrations/parse-integrations', () => {
-  return {
-    parseIntegrationsFile: (...args: unknown[]) => mockParseIntegrationsFile(...args),
-    getDefaultIntegrationsFilePath: (dir: string) => join(dir, DEFAULT_INTEGRATIONS_FILE),
-  }
-})
 // Mock fetchIntegrations for API integration tests
 const mockFetchIntegrations =
   vi.fn<(baseUrl: string, token: string, integrationIds?: string[]) => Promise<ApiIntegration[]>>()
-vi.mock('../integrations/fetch-integrations', async importOriginal => {
-  const actual = await importOriginal<typeof import('../integrations/fetch-integrations')>()
+vi.mock('@deepnote/database-integrations', async importOriginal => {
+  const actual = await importOriginal<typeof import('@deepnote/database-integrations')>()
   return {
     ...actual,
     fetchIntegrations: (...args: Parameters<typeof actual.fetchIntegrations>) => mockFetchIntegrations(...args),
@@ -75,7 +72,9 @@ vi.mock('../integrations/fetch-integrations', async importOriginal => {
 // Mock injectIntegrationEnvVars for testing integration env var injection
 const mockInjectIntegrationEnvVars = vi.fn<(integrations: unknown[], workingDirectory: string) => string[]>()
 mockInjectIntegrationEnvVars.mockReturnValue([])
-vi.mock('../integrations/inject-integration-env-vars', () => ({
+vi.mock('@deepnote/database-integrations/node', () => ({
+  parseIntegrationsFile: (...args: unknown[]) => mockParseIntegrationsFile(...args),
+  getDefaultIntegrationsFilePath: (dir: string) => join(dir, DEFAULT_INTEGRATIONS_FILE),
   injectIntegrationEnvVars: (...args: [unknown[], string]) => mockInjectIntegrationEnvVars(...args),
 }))
