--- packages/cli/src/constants.ts
@@ -1,8 +1,0 @@
-// Integration-related constants now live in @deepnote/database-integrations so they can
-// be shared with the VS Code extension. Re-exported here for backward compatibility.
-export {
-  BUILTIN_INTEGRATIONS,
-  DEEPNOTE_TOKEN_ENV,
-  DEFAULT_ENV_FILE,
-  DEFAULT_INTEGRATIONS_FILE,
-} from '@deepnote/database-integrations'
