--- packages/cli/src/integrations/fetch-integrations.ts
@@ -1,9 +1,0 @@
-// Cloud integration fetching now lives in @deepnote/database-integrations so it can be
-// reused by the VS Code extension. Re-exported here for backward compatibility.
-export {
-  type ApiIntegration,
-  type ApiResponse,
-  apiIntegrationSchema,
-  apiResponseSchema,
-  fetchIntegrations,
-} from '@deepnote/database-integrations'
