--- packages/cli/src/integrations/inject-integration-env-vars.ts
@@ -1,3 +1,0 @@
-// Injecting integration env vars into process.env now lives in
-// @deepnote/database-integrations/node. Re-exported here for backward compatibility.
-export { injectIntegrationEnvVars } from '@deepnote/database-integrations/node'
