--- packages/cli/src/integrations/integrations-file-schemas.ts
@@ -1,8 +1,0 @@
-// Integration file schemas now live in @deepnote/database-integrations. Re-exported here
-// for backward compatibility.
-export {
-  type BaseIntegrationsFile,
-  baseIntegrationsFileSchema,
-  type IntegrationsFile,
-  integrationsFileSchema,
-} from '@deepnote/database-integrations'
