--- packages/cli/src/utils/api.ts
@@ -1,2 +1,0 @@
-// ApiError now lives in @deepnote/database-integrations. Re-exported for backward compatibility.
-export { ApiError } from '@deepnote/database-integrations'
